tweak: move JsonRpcClient to its own dir
