+++ packages/api/src/json-rpc/index.ts
@@ -1,0 +1,1 @@
+export * from './JsonRpcClient.js';
